Allow mods to override ability icon colours (#400)
This change adds a copy of WOTC's `UITacticalHUD_Ability` class with slightly
modified code for rendering the colour. Specifically, the new code (based on
the old LW2 changes) fires an event that allow listeners to override the
ability colour using the provided XComLWTuple.

From a mod's perspective, it can now register listeners for the following
event:

 * 'OverrideAbilityIconColor'

   The event includes a single-element `XComLWTuple` of the form
   [ in bool IsObjective, out string Color]. The Color element has the
   default value for the corresponding ability's HUD icon. If you don't
   want to change it, leave the value as it is.

   The boolean lets you know whether the ability is associated with the
   current mission's objective. For example, Evac is an objective
   ability on Extract VIP missions.

   Lastly, the event source is the `XComGameState_Ability` instance for
   the icon that's being drawn.