Allow mods to override displayed ranks (#408)
Resolves issue #408. This commit introduces the following methods to
`XComGameState_Unit` that trigger events that listeners can use to
override rank names and icons:

 - GetSoldierRankName()
 - GetSoldierShortRankName()
 - GetSoldierRankIcon()

The commit also updates all the bits of code (outside of TLE and
multiplayer) that display soldier ranks and their icons so that they use
the above methods.

Here are the events that are fired:

 - `{ID: SoldierRankName, Data: [in int Rank, out string RankName]}`
 - `{ID: SoldierShortRankName, Data: [in int Rank, out string RankName]}`
 - `{ID: SoldierRankIcon, Data: [in int Rank, out string IconPath]}`

Note that the rank sent in the event may be -1, in which case the unit's
current rank is being requested.

The overall implementation is very similar to that of #106.