Allow mods to customise the reinforcements alert (#449)
This change modifies `UITacticalHUD_Countdown` so that it fires an event and
uses the response to determine whether to show the reinforcements alert, and
if so, what text (title and body) and colour to use.

The event takes the form of:

```
{
  ID: OverrideReinforcementsAlert,
  Data: [
    out bool ForceDisplay,
    out string Title,
    out string Body,
    out string Color],
}
```

`UIObjectiveList` also needed an update so that it wouldn't overlap the alert
when the latter was forced to be displayed by a listener. It does this by
checking with the listeners if they have forced the display of the alert or
not.