Allow mods to set anchor point for patrol zones (#500)
This change introduces a new 'OverrideEncounterZoneAnchorPoint' event
that allows mods to change the anchor point that XCOM 2 uses to
determine the patrol zones for pods. By default, the anchor point is
"dynamic" as it is the current location of the XCOM squad, which
typically changes throughout the mission.

The event takes the form:
```
{
   ID: OverrideEncounterZoneAnchorPoint,
   Data: [inout float X, inout float Y, inout float Z],
   Source: AIGroup
}
```
where X, Y and Z are the elements of the anchor point Vector.