Add event for overriding sound ranges (#510)
There is now a new event - 'OverrideSoundRange' - that allows listeners to
override the sound range for a weapon/ability combo. For example, it could
switch to using the sound range of the weapon's ammo or modify the existing
sound range based on weapon attachments or the unit's abilities.

Note that this is different from an earlier patch that added a hook for
customising an item's sound range. This adds to that by allowing mods to
take account of abilities, a unit's status, and so on.

The event itself takes the form:
```
  {
     ID: OverrideSoundRange,
     Data: [in XCGS_Unit SourceUnit, in XCGS_Item Weapon,
            in XCGS_Ability Ability, inout int SoundRange],
     Source: Unit (the one hearing the sound)
  }
```