Fix the disappearing rescue rings bug (#551)
In Gather Survivors and several LW2 missions, XCOM has to rescue units and
take them to evac. To enable player control of the rescued unit, its team
is swapped to XCom. But XGUnit then removes the range indicators from all
members of that unit's squad, which is not XCOM at that stage but team
Neutral.

This fix ensures that the swap team action changes the unit's squad (and
the pawn's team) in addition to the existing XGUnit team change.