Allow mods to override body/loot recovery (#571)
There are two new events, 'OverrideBodyRecovery' and
'OverrideLootRecovery', that allow mods to override the default
behaviour of whether incapacitated/dead soldiers and/or loot are
automatically recovered at the end of a mission.

The default behaviour is for this to happen only when a mission has
tactical objectives and those objectives have been completed.

The events take the form:
```
 {
   ID: OverrideBodyRecovery,
   Data: [inout bool DoBodyRecovery],
   Source: XComGameState_BattleData
 }
```
and
```
 {
   ID: OverrideLootRecovery,
   Data: [inout bool DoLootRecovery],
   Source: XComGameState_BattleData
 }
```