Address CHL Issue #533
This addresses the issue of allowing mods to override how the UISoldierHeader displays its stats. It uses 2 new classes that assist with this. The first is XComSolidierHeaderElement which is an object that holds information about a particular stat to display in the header like should the stat's max value be shown and should the value be colored. The other obect is XComSoldierHeader which is just a class that holds an array of XComSoldierHeaderElement objects and also a few referrences to other objects such as the Unit the stats are being displayed for and other items used to assist with stat calculation.

The XComSoldierHeader object will be passed to the 'OverrideSoldierHeader' event where the event listener can rearrange stats in any way and also add new stats if so desired. This also takes all of the stat calculation found in the PopulateData() function in UISoldierHeader and moves to a function called CalcAndDisplaySoldierStats() calculates stats and removes the call to SetSoldierStats(). SetSoldierStats has been left in the UISoldierHeader class in case other mods still make a call to it.