Allow mods to override the science score (#626)
I have added an 'OverrideScienceScore' event that's fired from
`XCGS_HeadquartersXCom.GetScienceScore()`. This event allows mods to
override the game's science score, for example to add their own bonuses
or to discount scientists that are doing other jobs.

The event takes the form:
```
 {
    ID: OverrideScienceScore,
    Data: [inout int ScienceScore, in bool AddLabBonus],
    Source: XCGS_HeadquartersXCom,
 }
```

Note that `AddLabBonus` indicates whether the game is requesting the
science score that is contributing to research progress (`true`) or is
checking whether there are enough scientists to meet a tech's or item's
science score requirement (`false`).

Resolves #626.