Fix teleporting enemy units basegame bug (#644)
Units in large pods could sometimes find themselves without any valid
tiles to walk through/to, which would cause them to teleport to another
part of the map.

This fix replicates what LW2 did to fix the same issue, by forcing some
valid tiles for the unit to move to. Thanks to Amineri (of Pavonis
Interactive) for the original code.

Fixes #644.