Block `eStatus_CovertAction` units from missions (#665)
Fixes `XCGS_Unit.CanGoOnMission()` to honour units with a status of
`eStatus_CovertAction`, returning `false` for those. This fixes a bug
where wounded SPARKs that are on a covert action (or some other activity
that sets that status) could be added to squads for missions via Squad
Select's autofill feature.

Fixes #665.