Fix covert action UI when rewards are prevented (#810)
The implementation of #438 allowed mods to block the giving of CA
rewards, but the UI still displayed that XP was gained and any popups
associated with the rewards (such as a scientist). In other words, you
could have a scientist reward blocked, but the player would still get
the popup saying they had gained a scientist!

`XComGameState_CovertAction` now records in a state variable whether the
rewards were prevented or not and the UI takes the value of that
variable into account.

I have also added documentation for the CovertAction_PreventGiveRewards
event.

Fixes #810.