Allow mods to change if a class can be trained (X2CommunityCore#814)
Resolves issue X2CommunityCore#814. This commit changes the UIChooseClass:GetClasses() function to introduce an event that validates if a class can be trained in the Guerilla Training School.

The event is fired once for each class whenever the GTS creates a list of classes to train. The listener simply needs to perform logic with the class name and return a boolean. It will return true by default, maintaining vanilla behavior if no listener changes it.

The event itself takes the form:
{
      ID: ValidateGTSClassTraining,
      Data: [in name SoldierClassName, out bool CanTrainClass],
      Source: self (UIChooseClass)
}