Allows mods to modify ability contexts (#890)
Abilities can have their own `ModifyNewContextFn` delegates that allow
modification of the ability context after it's created. However, it's
difficult for mods to override this in a compatible way, particularly if
multiple mods want to modify the same ability context.

This change introduces a 'PostModifyNewAbilityContext' event that's
fired after the `ModifyNewContextFn` delegate is called, allowing mods
to modify the ability context further.