Add extra damage modifier hooks for ApplyWeaponDamage (#923)
The existing API for weapon damage effects is problematic for
percentage-based damage modifiers because it makes them dependent on the
order in which effects are applied to a unit.

This change adds four extra hooks to `X2Effect_Persistent` that are used
by `CalculateDamageAmount()` and `GetDamagePreview()` in
`X2Effect_ApplyWeaponDamage` so that mods can apply modifiers that are
not dependent on effect order:

 * GetPreDefaultAttackingDamageModifier_CH()
 * GetPostDefaultAttackingDamageModifier_CH()
 * GetPreDefaultDefendingDamageModifier_CH()
 * GetPostDefaultDefendingDamageModifier_CH()

Note that these don't have to be used just for multiplicative,
percent-based multipliers. That is just the original motivation for
adding the hooks.