Fix duplicate psi ability training choices (#924)
A bug in `UIChoosePsiAbility.GetAbilities()` meant that under certain
circumstances, the screen could display the same ability twice, giving
the player only one other option to train.

The problem was that the local `AbilityName` variable was being
overridden by a new ability name before the code decided whether that
override would be displayed as a choice or not. This fix simply stores
each prerequisite ability name in a separate variable and only overrides
`AbilityName` if that prerequisite is actually replacing the original
ability as a choice.

Fixes #924.