Allow mods to override starting region links rule (#774)
`XComGameState_RegionLink.IsEligibleStartRegion()` now fires an
'OverrideAllowStartingRegionLink' event that allows mods to override the
rule for whether a region can be linked to a potential starting region
or not. The same event is used to bring `RemoveNonContinentLinks()` and
`XComGameState_WorldRegion.CanBeStartingRegion()` into line with
`IsEligibleStartRegion()`.

The event takes the form:
```
  EventID: OverrideAllowStartingRegionLink,
  EventData: [ in XComGameState_WorldRegion LinkedRegion, inout bool AllowLink ],
  EventSource: XComGameState_WorldRegion (potential starting region),
  NewGameState: none
```

Resolves #774.