Fix smoke and poison grenade targeting and effects
Grenades that apply effects to the world, such as smoke grenades, are
unreliable in their targeting because those effects are not always
applied to the tiles highlighted by the targeting. This change adds two
boolean config vars that, when true, fix this inconsistency so that
smoke and/or poison are applied to all the highlighted tiles.

In addition, there are some instances where the targeting will highlight
units without highlighting the tiles that they're on. This gives the
impression that those units will be affected by the world effects (like
smoke) when they won't. This changes adds a config array of grenade
names, where any grenade in that array will not highlight units unless
they are on highlighted tiles. Smoke grenades and bombs are the main
candidates for this array, but mods may add other types.

Resolves #669