Add `SetPriorityTargetStack` behavior tree node (#1009)
This adds a new behavior tree node that is based on `SetTargetStack`
except it *only* selects priority targets. It has no fallback to other
targets like `SetTargetStack` does.

It is primarily needed for AI that should attack priority targets, but
no others, for example units on yellow alert. Note that the AI config
needs to be modified to take advantage of this node; it doesn't do
anything by itself.

Resolves #1009.