Fix pods not scampering (#1008)
The `SkipTurn()` implementation in `XGAIBehavior` skips the entire pod's
turn if the pod starts the turn "unrevealed", which basically means they
haven't activated by the start of the turn. This happens when the pod
leader skips moving on activation, for example if they're immobilized or
panicked.

The fix I've implemented here is to check whether the pod is scampering
and only skip the pod's turn if they *aren't*. This ensures that the pod
members that can scamper actually do so.

Fixes #1008.