Add gameplay visibility condition to second shots (#1022)
The Rapid Fire and Chain Shot second shot abilities did not have the
gameplay visibility condition, so their valid targets included every
enemy unit on the map, which was causing performance issues in LWOTC.
There appears to be no reason for these abilities not to have this
condition, so adding them in.

Fixes #1022.