Modified UIAlert to use the "TrainingComplete" alert for positive traits - This has the same set of UI elements as Alert_NegativeSoldierEvent for negative traits but with a neutral colour scheme & promotion-like-icon.
Updated XcomHQPresentationLayer to check whether traits are positive or negative
before calling the notifybanner (this is the piece which places the trait notifications
on the bottom left of the screen). Adjusted the icons in the notification area to make
more sense.

Adjusted the UIAvengerHUD OnClickTraitIcon function so if someone clicks on a soldier
with a negative or positive trait, it takes them to the correct UI Panel.