Allow mods to override cost strings (#1218)
Adds the event `OverrideStrategyCostString` that allows mods to override
the cost strings created by the utility function
`UIUtilities_Strategy.GetStrategyCostString` which is used throughout
XCOM 2 to create the strings specifying the artifacts and resources
costs for items, facilities, covert actions, etc.

The event takes the form:
```event
EventID: OverrideStrategyCostString,
EventData: [
	in name ItemTemplateName,
	in int Quantity,
	in bool IsResourceCost,
	inout string CostString
],
EventSource: none,
NewGameState: none
```

The event is triggered individually for every component of the
`StrategyCost` value. Both the name of the item template of the cost
component and its are passed to the event. Futhermore, the information
whether it constitutes a resource cost or an artifact cost is passed via
`IsResourceCost`.

Resolves #1218